requirements.txt
curl_cffi>=0.7.0
groq>=0.11.0
python-dotenv>=1.0.0
